+++ tools/schema-design/src/widgets/components/FormRender/index.less
@@ -7,40 +7,3 @@
     width: 100%;
   }
 }
-
-// .fone-pro-form-modal {
-//   padding: 0 !important;
-// }
-
-// .fone-site-input-left {
-//   border-right-width: 0 !important;
-// }
-// .fone-site-input-left:hover,
-// .fone-site-input-left:focus {
-//   border-right-width: 1px;
-// }
-
-// .fone-site-input-right {
-//   border-left-width: 0;
-// }
-
-// .fone-site-input-right:hover,
-// .fone-site-input-right:focus {
-//   border-left-width: 1px;
-// }
-
-// .fone-site-field-section {
-//   border: 1px solid #d9d9d9;
-//   padding: 0 10px;
-//   line-height: 32px;
-//   border-left: none;
-//   border-right: none;
-//   background: #fff;
-//   color: #d9d9d9;
-// }
-
-// .ant-form-item-has-error {
-//   .fone-site-field-section {
-//     border: 1px solid #ff4d4f;
-//   }
-// }
